# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Instituto Acaz Ribeiro - Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List session contracts
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: professionalId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a session contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}/sessions:
    get:
      operationId: listContractSessions
      tags: [contracts]
      summary: List sessions for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractSession"
    post:
      operationId: addContractSession
      tags: [contracts]
      summary: Record a session for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSessionBody"
      responses:
        "201":
          description: Session added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSession"

  # Dashboard
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month]
            default: today
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /dashboard/recent-sales:
    get:
      operationId: getRecentSales
      tags: [dashboard]
      summary: Get recent sales
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
  /dashboard/professional-stats:
    get:
      operationId: getProfessionalStats
      tags: [dashboard]
      summary: Get stats per professional (for current user if professional role)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Professional stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfessionalStat"

  # Reports
...
components:
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        professionalId:
          type: integer
        professionalName:
          type: string
        serviceId:
          type: ["integer", "null"]
        serviceName:
          type: ["string", "null"]
        packageId:
          type: ["integer", "null"]
        packageName:
          type: ["string", "null"]
        grossAmount:
          type: number
        paymentMethod:
          type: string
          enum: [pix, credit, debit, cash]
        gateFee:
          type: number
          description: Payment gateway fee
        netAmount:
          type: number
          description: Gross minus gateway fee
        clinicAmount:
          type: number
          description: Clinic's share
        professionalAmount:
          type: number
          description: Professional's share
        appointmentId:
          type: ["integer", "null"]
        contractId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - professionalId
        - professionalName
        - serviceId
        - serviceName
        - packageId
        - packageName
        - grossAmount
        - paymentMethod
        - gateFee
        - netAmount
        - clinicAmount
        - professionalAmount
        - appointmentId
        - contractId
        - notes
        - createdAt
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        professionalId:
          type: integer
        professionalName:
          type: string
        packageId:
          type: integer
        packageName:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        totalSessions:
          type: integer
        usedSessions:
          type: integer
        remainingSessions:
          type: integer
        saleId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, completed, cancelled]
        expiresAt:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - professionalId
        - professionalName
        - packageId
        - packageName
        - serviceId
        - serviceName
        - totalSessions
        - usedSessions
        - remainingSessions
        - saleId
        - status
        - expiresAt
        - notes
        - createdAt
    CreateContractBody:
      type: object
      properties:
        clientId:
          type: integer
        professionalId:
          type: integer
        packageId:
          type: integer
        saleId:
          type: integer
        notes:
          type: string
      required:
        - clientId
        - professionalId
        - packageId
    UpdateContractBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, cancelled]
        notes:
          type: string
    ContractSession:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        sessionNumber:
          type: integer
        performedAt:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
      required:
        - id
        - contractId
        - appointmentId
        - sessionNumber
        - performedAt
        - notes
    AddSessionBody:
      type: object
      properties:
        appointmentId:
          type: integer
        performedAt:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - performedAt

    # Dashboard
    DashboardSummary:
      type: object
      properties:
        totalAppointments:
          type: integer
        completedAppointments:
          type: integer
        cancelledAppointments:
          type: integer
        totalRevenue:
          type: number
        clinicRevenue:
          type: number
        professionalRevenue:
          type: number
        newClients:
          type: integer
        activeContracts:
          type: integer
      required:
        - totalAppointments
        - completedAppointments
        - cancelledAppointments
        - totalRevenue
        - clinicRevenue
        - professionalRevenue
        - newClients
        - activeContracts
    ProfessionalStat:
      type: object
      properties:
        professionalId:
          type: integer
        professionalName:
          type: string
        appointments:
          type: integer
        revenue:
          type: number
        professionalEarnings:
          type: number
        clinicEarnings:
          type: number
      required:
        - professionalId
        - professionalName
        - appointments
        - revenue
        - professionalEarnings
        - clinicEarnings

    # Reports
...